implement confirm order none implement is receive receipt and time formattat not correct

+++ app/src/main/java/com/example/rollinginthebeef/modules/Order.kt
@@ -17,7 +17,7 @@ class Order(
     @SerializedName("order_date") val order_date: String,
 
     @Expose
-    @SerializedName("order_total") val order_total: Float,
+    @SerializedName("order_total") val order_total: Int,
 
     @Expose
     @SerializedName("order_status") val order_status: Int) {}
